# Conflicts:
#	juntagrico/admins/job_admin.py
#	juntagrico/admins/one_time_job_admin.py
#	settings/minimal.py

+++ juntagrico/entity/subtypes.py
@@ -66,10 +66,8 @@ class SubscriptionType(JuntagricoBaseModel):
                              related_name='types', verbose_name=_('Grösse'))
     shares = models.PositiveIntegerField(
         _('Anz benötigter Anteilsscheine'), default=0)
-    required_assignments = models.PositiveIntegerField(
-        _('Anz benötigter Arbeitseinsätze'))
-    required_core_assignments = models.PositiveIntegerField(
-        _('Anz benötigter Kern Arbeitseinsätze'), default=0)
+    required_assignments = models.FloatField(_('Anz benötigter Arbeitseinsätze'))
+    required_core_assignments = models.FloatField(_('Anz benötigter Kern Arbeitseinsätze'), default=0)
     price = models.DecimalField(_('Preis'), max_digits=9, decimal_places=2)
     visible = models.BooleanField(_('Sichtbar'), default=True)
     trial = models.BooleanField(_('Probe-Abo'), default=False)  # Deprecated
